Refactor Prayer Request Actions (Draft)
